Menu - permission module added

+++ resources/views/permissions/index.blade.php
@@ -1,21 +1,19 @@
 @extends('layouts.app-master')
 
 @section('content')
-    
-    <h1 class="mb-3">Laravel 8 User Roles and Permissions Step by Step Tutorial - codeanddeploy.com</h1>
+
+<div class="content">
+    <div class="container-fluid">   
+     
+        @include('layouts.partials.messages')
 
     <div class="bg-light p-4 rounded">
-        <h2>Permissions</h2>
-        <div class="lead">
-            Manage your permissions here.
-            <a href="{{ route('permissions.create') }}" class="btn btn-primary btn-sm float-right">Add permissions</a>
-        </div>
-        
-        <div class="mt-2">
-            @include('layouts.partials.messages')
-        </div>
-
-        <table class="table table-striped">
+        <h2>Menus & Permissions <a href="{{ route('permissions.create') }}" class="btn btn-primary btn-sm float-right">Add Menu</a></h2>
+        <h2 class="small">
+            Manage your permissions and menus here.
+        </h2>
+      
+        <table id="datatables" class="table table-striped" cellspacing="0" width="100%" style="width:100%">
             <thead>
             <tr>
                 <th scope="col">Menu Name</th>
@@ -24,7 +22,7 @@
                 <th scope="col">Sort Order</th>
                 <th scope="col">Menu Icon</th>
                 <th scope="col">Guard</th> 
-                <th scope="col" colspan="3" width="1%"></th> 
+                <th scope="col"></th> 
             </tr>
             </thead>
             <tbody>
@@ -36,9 +34,9 @@
                         <td>{{ $permission->sort_order }}</td>
                         <td>{{ $permission->menu_icon }}</td>
                         <td>{{ $permission->guard_name }}</td>
-                        <td><a href="{{ route('permissions.edit', $permission->id) }}" class="btn btn-info btn-sm">Edit</a></td>
-                        <td>
-                            {!! Form::open(['method' => 'DELETE','route' => ['permissions.destroy', $permission->id],'style'=>'display:inline']) !!}
+                        <td><a href="{{ route('permissions.edit', $permission->id) }}" class="btn btn-info btn-sm">Edit</a>
+                        
+                            {!! Form::open(['method' => 'DELETE','route' => ['permissions.destroy', $permission->id],'style'=>'display:inline', 'onClick' => 'return confirm("Are you sure you want to delete ?");']) !!}
                             {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
                             {!! Form::close() !!}
                         </td>
@@ -48,4 +46,8 @@
         </table>
 
     </div>
+
+
+</div>
+</div>
 @endsection
